refactor: 1.basic 좀더 컴포넌트를 분리 시킴. 2. 네이밍도 같이 수정
