+++ src/basic/components/shop/checkout/cart/EmptyCart.tsx
@@ -1,4 +1,4 @@
-import { EmptyCartIcon } from '../icons';
+import { EmptyCartIcon } from '../../../icons';
 
 const EmptyCart = () => {
   return (
